expanded use for higher latent dimensions and smaller crop sizes improved detection by filters countours by area

+++ eval/ae_eval.py
@@ -4,25 +4,24 @@ import cv2
 import matplotlib.pyplot as plt
 import numpy as np
 import tensorflow as tf
-from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, classification_report
+from sklearn.metrics import ConfusionMatrixDisplay, confusion_matrix, classification_report
 from sklearn.metrics import precision_recall_curve, average_precision_score, PrecisionRecallDisplay
 from tqdm import tqdm
 
 from dataloader.image_generators import test_image_generator, train_val_image_generator
 from models.autoencoders import CAE
-from train.losses import mse_ssim_mixed
 
 physical_devices = tf.config.experimental.list_physical_devices('GPU')
 if len(physical_devices) > 0:
     tf.config.experimental.set_memory_growth(physical_devices[0], True)
 
 
-def eval_cae_anomaly_scores(path_to_data, path_to_weights, batch_size=1, first_run=True):
+def eval_cae_anomaly_scores(path_to_data, path_to_weights, metric_fn, batch_size=1, crop_size=128, latent_dim=256,
+                            first_run=True):
     if first_run:
-        crop_size = 256
-        test_img_gen = test_image_generator(path_to_data, batch_size=1, crop_size=crop_size)
+        test_img_gen = test_image_generator(path_to_data, batch_size=batch_size, crop_size=crop_size)
 
-        cae = CAE(input_shape=(crop_size, crop_size, 1))
+        cae = CAE(input_shape=(crop_size, crop_size, 1), latent_dim=latent_dim)
 
         if os.path.exists(path_to_weights):
             cae.load_weights(path_to_weights)
@@ -31,31 +30,21 @@ def eval_cae_anomaly_scores(path_to_data, path_to_weights, batch_size=1, first_r
         labels_total = []
         for cogwheel_crops, labels in tqdm(test_img_gen):
             cogwheel_crops = tf.convert_to_tensor(cogwheel_crops, dtype=tf.float32)
-            anomaly_scores = cae.anomaly_scores(cogwheel_crops, metric_fn=mse_ssim_mixed)
-            anomaly_scores_total.append(anomaly_scores)
+            anomaly_scores = cae.anomaly_scores(cogwheel_crops, metric_fn=metric_fn)
+            anomaly_scores_total.extend(anomaly_scores)
             labels_total.extend(labels)
-
-        anomaly_scores_total = np.stack(anomaly_scores_total, axis=0)
-        labels_total = np.stack(labels_total, axis=0)
-
-        np.savez('eval_results.npz', anomaly_scores=anomaly_scores_total, labels=labels_total)
-
-    else:
-        saved_data = np.load('eval_results.npz')
-        anomaly_scores_total = saved_data['anomaly_scores']
-        labels_total = saved_data['labels']
 
     return anomaly_scores_total, labels_total
 
 
-def eval_cae_detect_anomalies_by_crop(path_to_defects_data, path_to_weights, median_training_loss,
-                                      preprocess=True, debug=False, first_run=True):
+def eval_cae_detect_anomalies_by_crop(path_to_defects_data, path_to_weights, min_area=9,
+                                      latent_dim=256, crop_size=128, batch_size=1,
+                                      preprocess=False, debug=False, first_run=True):
     if first_run:
-        crop_size = 256
         test_img_gen = test_image_generator(path_to_defects_data, batch_size=1, crop_size=crop_size,
-                                            preprocess=preprocess,)
+                                            preprocess=preprocess, )
 
-        cae = CAE(input_shape=(crop_size, crop_size, 1))
+        cae = CAE(input_shape=(crop_size, crop_size, 1), latent_dim=latent_dim)
 
         if os.path.exists(path_to_weights):
             cae.load_weights(path_to_weights)
@@ -65,16 +54,16 @@ def eval_cae_detect_anomalies_by_crop(path_to_defects_data, path_to_weights, med
         for cogwheel_crop, label in tqdm(test_img_gen):
             labels.extend(label)
             crop_for_display = cogwheel_crop[0].astype(np.uint8)
-            bboxes = cae.detect_anomalies(cogwheel_crop, mean_loss=median_training_loss, label=label, debug=debug)
+            bboxes = cae.detect_anomalies(cogwheel_crop, label=label, debug=debug, crop_size=crop_size,
+                                          min_area=min_area)
             pred = (len(bboxes) > 0)
             predictions.append(pred)
             if pred:
-                if debug:
+                if debug and label == 1:
                     cv2.drawContours(crop_for_display, bboxes, -1, (255, 0, 0), 3)
                     cv2.imshow(f'GT Label - {str(label)}', crop_for_display)
                     cv2.waitKey(0)
                     cv2.destroyAllWindows()
-
 
         predictions = np.stack(predictions, axis=0)
         labels = np.stack(labels, axis=0)
@@ -82,55 +71,6 @@ def eval_cae_detect_anomalies_by_crop(path_to_defects_data, path_to_weights, med
 
     else:
         saved_data = np.load('det_eval_results.npz')
-        predictions = saved_data['predictions']
-        labels = saved_data['labels']
-
-    return predictions, labels
-
-
-def eval_cae_detect_anomalies_by_images(path_to_normal_data, path_to_defect_data, path_to_weights,
-                                        debug=False,
-                                        first_run=True):
-    if first_run:
-        crop_size = 256
-        defect_img_gen = test_image_generator(path_to_defect_data, batch_size=64, crop_size=crop_size)
-        normal_img_gen = train_image_generator(path_to_normal_data, batch_size=64, crop_size=crop_size, vrepeat=False)
-
-        cae = CAE(input_shape=(crop_size, crop_size, 1))
-
-        if os.path.exists(path_to_weights):
-            cae.load_weights(path_to_weights)
-
-        agg_labels = []
-        agg_predictions = []
-        for cogwheel_crops, _ in tqdm(defect_img_gen):
-            agg_labels.append(1)
-            pred = 0
-            for crop in cogwheel_crops:
-                crop = np.expand_dims(crop, axis=0)
-                bboxes = cae.detect_anomalies(crop, debug=debug)
-                if len(bboxes) > 0:
-                    pred = 1
-                    break
-            agg_predictions.append(pred)
-
-        for cogwheel_crops, _ in tqdm(normal_img_gen):
-            agg_labels.append(0)
-            pred = 0
-            for crop in cogwheel_crops:
-                crop = np.expand_dims(crop, axis=0)
-                bboxes = cae.detect_anomalies(crop, debug=debug)
-                if len(bboxes) > 0:
-                    pred = 1
-                    break
-            agg_predictions.append(pred)
-
-        predictions = np.stack(agg_predictions, axis=0)
-        labels = np.stack(agg_labels, axis=0)
-        np.savez('det_by_image_eval_results.npz', predictions=predictions, labels=labels)
-
-    else:
-        saved_data = np.load('det_by_image_eval_results.npz')
         predictions = saved_data['predictions']
         labels = saved_data['labels']
 
@@ -145,25 +85,25 @@ def save_precision_recall_curve(anomaly_scores, labels):
     plt.savefig('precision_recall_curve.png', dpi=400)
 
 
-def visualize_diff(path_to_images, path_to_weights, method='heatmap'):
-    crop_size = 256
-    test_img_gen = test_image_generator(path_to_images, batch_size=64, crop_size=crop_size)
+def visualize_diff(path_to_images, path_to_weights, method='heatmap', latent_dim=256, crop_size=128, batch_size=128):
+    test_img_gen = test_image_generator(path_to_images, batch_size=128, crop_size=crop_size, preprocess=False)
 
-    cae = CAE(input_shape=(crop_size, crop_size, 1))
+    cae = CAE(input_shape=(crop_size, crop_size, 1), latent_dim=latent_dim)
 
     if os.path.exists(path_to_weights):
         cae.load_weights(path_to_weights)
 
     for cogwheel_crops, labels in test_img_gen:
-        cae.visualize_anomalies(cogwheel_crops, method=method, labels=labels)
+        cae.visualize_anomalies(cogwheel_crops, method=method, labels=labels, crop_size=crop_size)
 
 
-def find_median_threshold(path_to_images, path_to_weights, metric_fn, sample_frac=None):
-    crop_size = 256
+def find_median_threshold(path_to_images, path_to_weights, metric_fn, crop_size=128, latent_dim=256,
+                          sample_frac=None,
+                          batch_size=1):
     img_gen, _ = train_val_image_generator(path_to_images, batch_size=1, val_ratio=0.0, crop_size=crop_size,
-                                                  repeat=False, sample_frac=sample_frac)
+                                           repeat=False, sample_frac=sample_frac)
 
-    cae = CAE(input_shape=(crop_size, crop_size, 1))
+    cae = CAE(input_shape=(crop_size, crop_size, 1), latent_dim=latent_dim)
 
     if os.path.exists(path_to_weights):
         cae.load_weights(path_to_weights)
@@ -185,28 +125,33 @@ def find_median_threshold(path_to_images, path_to_weights, metric_fn, sample_fra
 def main():
     defect_data_path = "/media/jpowell/hdd/Data/AIS/RO2_NG_images/"
     normal_data_path = "/media/jpowell/hdd/Data/AIS/RO2_OK_images/"
-    path_to_weights = '/train/weights/no_preprocess_best_weights.h5'
+    path_to_weights = '/home/jpowell/PycharmProjects/AIS/ais_aae/train/128x_256d_best_model.h5'
+
+    nn_params = {
+        'crop_size': 128,
+        'latent_dim': 256,
+        'batch_size': 128
+    }
 
     # median_threshold = find_median_threshold(normal_data_path, path_to_weights, mse_ssim_mixed,
-    #                                          sample_frac=0.1)
+    #                                          sample_frac=0.01, **nn_params)
     # print(median_threshold)
 
-    # visualize_diff(defect_data_path, path_to_weights, method='lpf')
+    # visualize_diff(defect_data_path, path_to_weights, method='triptych', **nn_params)
 
     # anomaly_scores, labels = eval_cae_anomaly_scores(defect_data_path, path_to_weights,
-    #                                                  first_run=True)
+    #                                                  metric_fn=mse_dssim_mixed,
+    #                                                  first_run=True,
+    #                                                  **nn_params)
     # save_precision_recall_curve(anomaly_scores, labels)
 
     predictions, labels = eval_cae_detect_anomalies_by_crop(defect_data_path,
                                                             path_to_weights,
-                                                            median_training_loss=80,
                                                             preprocess=False,
                                                             first_run=True,
-                                                            debug=False)
-
-
-    # predictions, labels = eval_cae_detect_anomalies_by_images(normal_data_path, defect_data_path, path_to_weights,
-    #                                                           first_run=True)
+                                                            debug=False,
+                                                            min_area=10,
+                                                            **nn_params)
 
     cm = confusion_matrix(labels, predictions)
     print(classification_report(labels, predictions, target_names=['Normal', 'Anomaly']))
